refactor(base-app): encapsulate text profile form into a module
