+++ apps/base-app/src/features/textProfileForm/ui/TextProfileForm.vue
@@ -1,10 +1,10 @@
 <script setup lang="ts">
 import { ref, computed } from 'vue'
 import { useI18n } from 'vue-i18n-composable'
-import { type Settings, getLangConfig } from '@readapt/settings'
+import { type Settings, type Language, getLangConfig } from '@readapt/settings'
 import BaseTabs from '@/shared/ui/BaseTabs.vue'
-import SettingsMenuGeneral from '@/views/SettingsMenuGeneral.vue'
-import SettingsMenuTableItems from '@/views/SettingsMenuTableItems.vue'
+import TextProfileFormGeneralTab from './TextProfileFormGeneralTab.vue'
+import TextProfileFormItemsTab from './TextProfileFormItemsTab.vue'
 
 interface Props {
   settings: Settings
@@ -17,25 +17,26 @@ const tabNames = [t('SETTINGS.GENERAL_SETTINGS'), t('SETTINGS.PHONEME_SETTINGS')
 const activeTabIndex = ref(0)
 
 interface Emits {
-  (event: 'change', value: unknown): void
+  (event: 'update-settings', value: unknown): void
+  (event: 'set-language', value: Language): void
 }
 const emit = defineEmits<Emits>()
 const emitUpdateSettings = (event: unknown) => emit('update-settings', event)
+const emitChangeLanguage = (language: Language) => emit('change-language', language)
 
 const languageConfig = computed(() => getLangConfig(props.settings.language))
 </script>
 <template>
   <div>
     <BaseTabs data-test-id="tabs" :names="tabNames" v-model="activeTabIndex" />
-    <SettingsMenuGeneral
+    <TextProfileFormGeneralTab
       v-if="activeTabIndex === 0"
       :settings="settings"
       data-test-id="general"
       @update="emitUpdateSettings($event)"
-      @change-language="$emit('change-language', $event)"
-    >
-    </SettingsMenuGeneral>
-    <SettingsMenuTableItems
+      @change-language="emitChangeLanguage($event)"
+    />
+    <TextProfileFormItemsTab
       v-if="activeTabIndex === 1"
       table-label="SETTINGS.PHONEME"
       switch-all-label="SETTINGS.ALL_PHONEMES_SETTINGS"
@@ -45,9 +46,8 @@ const languageConfig = computed(() => getLangConfig(props.settings.language))
       :options="languageConfig.phonemeOptions"
       @update-items="emitUpdateSettings({ key: 'phonemes', value: $event })"
       @update-active="emitUpdateSettings({ key: 'phonemesActive', value: $event })"
-    >
-    </SettingsMenuTableItems>
-    <SettingsMenuTableItems
+    />
+    <TextProfileFormItemsTab
       v-if="activeTabIndex === 2"
       table-label="SETTINGS.LETTER"
       switch-all-label="SETTINGS.ALL_LETTERS_SETTINGS"
@@ -57,6 +57,6 @@ const languageConfig = computed(() => getLangConfig(props.settings.language))
       :options="languageConfig.letterOptions"
       @update-items="emitUpdateSettings({ key: 'letters', value: $event })"
       @update-active="emitUpdateSettings({ key: 'lettersActive', value: $event })"
-    ></SettingsMenuTableItems>
+    />
   </div>
 </template>
