+++ apps/base-app/src/features/textProfileForm/ui/TextProfileFormItemsTab.vue
