+++ apps/base-app/src/pages/options/model/useFormSettings.ts
@@ -10,7 +10,8 @@ export function useFormSettings(selectedProfileId: Ref<TextProfileId | null>) {
 
   const settings = computed(() => settingsByLanguage.value[language.value])
   const changeLanguage = (newLanguage: Language) => (language.value = newLanguage)
-  const updateSettings = <K extends keyof Settings>(key: K, value: Settings[K]) => {
+
+  const updateSettings = <K extends keyof Settings>({ key, value }: { key: K; value: Settings[K] }) => {
     settings.value[key] = value
   }
